refactored most functionality into KnownHost class
Most of the code that was in Form1 didn't really belong there, so it's been
moved into the KnownHost class.

KnownHost also, perhaps questionably, instantiates its own ListViewItem, and
opens a handle into the Windows Registry.

This comes at some advantage, though, allowing Form1 to focus on the rest of the
GUI. Perhaps the GUI-related functions of KnownHost can be broken out into its
own thing.

The next step hopefully is to allow KnownHost to be instantiated from data we
read from a .reg file, and for the KnownHost to know whether or not the
host+publickey that it represents is currently in the Registry or not.