fixed regression in fingerprint calculation
Earlier I "fixed" a bug with calculating hashes of keys from openssh at the cost
of breaking hash calculation for putty's keys stored in windows registry.

It turns out that I wasn't imagining things: OpenSSH really does NUL pad the
modulus of RSA keys. It appears to be a behavior of OpenSSL and doesn't seem to
be documented in RFC 4716 or RFC 4253 as near as I can tell (though I just
skimmed them and maybe it's outside the scope of those two documents.)